requirements.txt
flask
flask-cors
spacy
sentence-transformers
PyMuPDF
PyGithub
radon
python-dotenv